fix: actually use Metal, and fix the kernel gaps that surfaced once it did
main.rs called Device::cuda_if_available(0), but Cargo.toml only enables
candle's "metal" feature (no "cuda") — that check was always false, so
every run silently fell back to CPU regardless of the Metal feature being
compiled in. Switched to Device::metal_if_available(0).

Once actually running on Metal, four latent kernel-support gaps surfaced
that the CPU backend had been silently tolerating:

- Tensor::randn(0.0f64, 1.0f64, ...) — Metal has no F64 rand_uniform
  kernel. Switched the two call sites to f32 literals.
- pipeline.rs's per-step timestep was cast to f64 then immediately back to
  f32 for no reason — Metal has no F64->F32 cast kernel either. Dropped
  the pointless f64 round-trip.
- rope.rs's Tensor::arange(0, 8192, device) used bare integer literals,
  which Rust defaults to i32 absent other constraints; the subsequent
  .to_dtype(F32) then hit the same missing-cast-kernel issue. Made the
  literals explicitly i64 (matching every other arange call site in the
  file, which already had an explicit cast forcing i64 inference).
- Attention's `v` stayed a transposed (non-contiguous) view from
  reshape+transpose all the way to the final attn_weights.matmul(&v);
  Metal's matmul kernel is apparently stricter about this than CPU's.
  Added .contiguous() right after the transpose, in both the quantized
  and non-quantized Attention::forward (and forward_cached, for when the
  KV-cache path is eventually wired up).

Verified: cargo test still passes (6 tests); the full pipeline now runs
end to end on Metal, confirmed via `Device: Metal(...)` in the output.
Denoising is roughly 11-15x faster than the same run on CPU (4 steps:
173s -> 11.9s; 20 steps: 869.8s -> 74.6s), with VAE decode dropping from
~6s to well under 1s. Output quality/artifact character (see the previous
commit's checkerboard-texture finding) is unchanged between CPU and Metal
runs at the same step count, which is further evidence that artifact is a
content/architecture issue rather than a backend-specific bug.

+++ src/pipeline.rs
@@ -68,7 +68,8 @@ pub fn denoise(
 
     // Prepare random latents: [B, C, T, H, W] (temporal at index 2)
     let shape = (batch_size, num_channels_latents, 1, latent_h, latent_w);
-    let latents = Tensor::randn(0.0f64, 1.0f64, shape, &cfg.device)?;
+    // f32, not f64: Metal has no F64 rand_uniform kernel.
+    let latents = Tensor::randn(0.0f32, 1.0f32, shape, &cfg.device)?;
     let mut packed_latents = pack_latents(&latents)?.to_dtype(cfg.dtype)?;
     let prompt_emb = prompt_emb.to_dtype(cfg.dtype)?;
 
@@ -93,9 +94,10 @@ pub fn denoise(
 
     // Denoising loop
     for i in 0..num_inference_steps {
-        let t = timesteps[i] as f64;
+        let t = timesteps[i];
         // [B] (batch=1); `timestep_embedding` does its own unsqueeze to broadcast
-        // against the exponent table, so this must stay 1-D here.
+        // against the exponent table, so this must stay 1-D here. f32, not f64:
+        // Metal has no F64->F32 dtype-cast kernel.
         let timestep = Tensor::new(&[t], &cfg.device)?
             .to_dtype(packed_latents.dtype())?;
